requirements.txt
Pillow>=11.0